requirements.txt
# Framework e Servidor
Flask==3.0.3
gunicorn==22.0.0

# Banco de Dados
psycopg2-binary==2.9.9

# Análise de Dados
pandas==2.2.2
openpyxl==3.1.5

# Upload e Nuvem
cloudinary==1.41.0

# PDF e Requisições Web
PyMuPDF==1.24.1
requests==2.32.3

# Utilidades
python-dateutil==2.9.0.post0
python-dotenv==1.0.1

Flask-Login==0.6.3
Werkzeug==3.1.3

Flask-SQLAlchemy==3.1.1
waitress==3.0.0